feat(scene): Adds a grabbable lantern object
Adds a lantern prefab to the scene, which is made grabbable.

The lantern object includes a Rigidbody, an InteractionBehaviour script, and a Grab_Anchor to facilitate interaction.

Also updates the URP global settings by re-ordering the serialized list IDs and applies a change to the collision detection of another prefab